my-pets - move edit dog actions and components into _lib
