+++ src/app/user/(logged-in)/my-pets/_lib/components/edit-dog-profile-form.tsx
@@ -5,9 +5,9 @@ import { DogProfile } from "@/lib/dog/dog-models";
 import { useRouter } from "next/navigation";
 import { RoutePath } from "@/lib/route-path";
 import { Err, Ok, Result } from "@/lib/utilities/result";
-import { postDogProfileUpdate } from "../_actions/post-dog-profile-update";
+import { postDogProfileUpdate } from "../actions/post-dog-profile-update";
 import { CODE } from "@/lib/utilities/bark-code";
-import { GeneralDogForm } from "../../../_lib/components/general-dog-form";
+import { GeneralDogForm } from "./general-dog-form";
 import { useToast } from "@/components/ui/use-toast";
 import { MINIMUM_TOAST_MILLIS } from "@/app/_lib/toast-delay";
 import { asyncSleep } from "@/lib/utilities/async-sleep";
